requirements.txt
streamlit
Pillow
pdf2image
pytesseract
PyPDF2
python-docx 
